requirements.txt
Django==4.2.3
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.3
click==7.1.2
decorator==4.4.2
django-bootstrap-form==3.4
django-filter==2.4.0
django-tinymce==3.1.0p
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
pytz==2020.1
self==0.0.0
soupsieve==2.0.1
sqlparse==0.4.1
Werkzeug==1.0.1
